projects/beat-the-bookies/data-process/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.9.0
async-generator==1.10
async-timeout==4.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.0
decorator==5.1.0
Flask==2.0.2
frozenlist==1.2.0
fsspec==2021.11.1
gcsfs==2021.11.1
google-api-core==2.3.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-core==2.2.1
google-cloud-pubsub==2.9.0
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
h11==0.12.0
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.6.0
libcst==0.3.23
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.2.0
mypy-extensions==0.4.3
numpy==1.21.4
oauthlib==3.1.1
outcome==1.1.0
pandas==1.3.5
platformdirs==2.4.0
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.0.1
typing-inspect==0.7.1
urllib3==1.26.7
webdriver-manager==3.5.2
Werkzeug==2.0.2
wrapt==1.13.3
wsproto==1.0.0
yarl==1.7.2
